Move HydeKernel into Foundation namespace
