+++ app/bootstrap.php
@@ -47,17 +47,17 @@ $app->singleton(
 |
 */
 
-$hyde = new Hyde\Framework\HydeKernel(
+$hyde = new \Hyde\Foundation\HydeKernel(
     dirname(__DIR__)
 );
 
 $app->singleton(
-    Hyde\Framework\HydeKernel::class, function () {
-        return Hyde\Framework\HydeKernel::getInstance();
+    \Hyde\Foundation\HydeKernel::class, function () {
+        return \Hyde\Foundation\HydeKernel::getInstance();
     }
 );
 
-Hyde\Framework\HydeKernel::setInstance($hyde);
+\Hyde\Foundation\HydeKernel::setInstance($hyde);
 
 /*
 |--------------------------------------------------------------------------
