+++ packages/framework/src/Foundation/Concerns/ForwardsFilesystem.php
@@ -7,7 +7,7 @@ namespace Hyde\Foundation\Concerns;
 /**
  * @internal Single-use trait for the HydeKernel class.
  *
- * @see \Hyde\Framework\HydeKernel
+ * @see \Hyde\Foundation\HydeKernel
  */
 trait ForwardsFilesystem
 {
