+++ packages/framework/src/Foundation/Concerns/HandlesFoundationCollections.php
@@ -11,7 +11,7 @@ use Hyde\Foundation\RouteCollection;
 /**
  * @internal Single-use trait for the HydeKernel class.
  *
- * @see \Hyde\Framework\HydeKernel
+ * @see \Hyde\Foundation\HydeKernel
  */
 trait HandlesFoundationCollections
 {
