+++ packages/framework/src/Foundation/Concerns/ManagesViewData.php
@@ -11,7 +11,7 @@ use Illuminate\Support\Facades\View;
 /**
  * @internal Single-use trait for the HydeKernel class.
  *
- * @see \Hyde\Framework\HydeKernel
+ * @see \Hyde\Foundation\HydeKernel
  */
 trait ManagesViewData
 {
