+++ packages/framework/src/Foundation/Filesystem.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Hyde\Foundation;
 
-use Hyde\Framework\HydeKernel;
 use Hyde\Framework\Models\Pages\BladePage;
 use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
