+++ packages/framework/src/Foundation/Hyperlinks.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Hyde\Foundation;
 
 use Hyde\Framework\Exceptions\BaseUrlNotSetException;
-use Hyde\Framework\HydeKernel;
 use Hyde\Framework\Models\Pages\DocumentationPage;
 
 /**
