+++ packages/framework/src/Hyde.php
@@ -5,18 +5,18 @@ declare(strict_types=1);
 namespace Hyde;
 
 use Hyde\Foundation\FileCollection;
+use Hyde\Foundation\HydeKernel;
 use Hyde\Foundation\PageCollection;
 use Hyde\Foundation\RouteCollection;
 use Hyde\Framework\Concerns\HydePage;
 use Hyde\Framework\Helpers\Features;
-use Hyde\Framework\HydeKernel;
 use Hyde\Framework\Models\Support\Route;
 use Illuminate\Support\Facades\Facade;
 
 /**
  * General facade for Hyde services.
  *
- * @see \Hyde\Framework\HydeKernel
+ * @see \Hyde\Foundation\HydeKernel
  *
  * @author  Caen De Silva <caen@desilva.se>
  * @copyright 2022 Caen De Silva
