+++ packages/framework/tests/Unit/HydeBasePathCanBeChangedTest.php
@@ -10,9 +10,9 @@ use Hyde\Testing\TestCase;
 /**
  * Class HydeBasePathCanBeChangedTest.
  *
- * @covers \Hyde\Framework\HydeKernel::getBasePath
- * @covers \Hyde\Framework\HydeKernel::setBasePath
- * @covers \Hyde\Framework\HydeKernel::path
+ * @covers \Hyde\Foundation\HydeKernel::getBasePath
+ * @covers \Hyde\Foundation\HydeKernel::setBasePath
+ * @covers \Hyde\Foundation\HydeKernel::path
  */
 class HydeBasePathCanBeChangedTest extends TestCase
 {
