+++ packages/framework/tests/Unit/HydeGetBasePathHasFallbackTest.php
@@ -10,7 +10,7 @@ use Hyde\Testing\TestCase;
 /**
  * Class HydeGetBasePathHasFallbackTest.
  *
- * @covers \Hyde\Framework\HydeKernel::getBasePath
+ * @covers \Hyde\Foundation\HydeKernel::getBasePath
  */
 class HydeGetBasePathHasFallbackTest extends TestCase
 {
